Removed php-imap extension from default list of installed php modules for better portability:
The imap extension is not needed by Magento. It was included likely due to this role having been used once upon a time for deployment of a ticketing system which used imap accounts. Since this is not needed by Magento and the imap extension is not available to us in EL 8 streams, it has been removed completely.